исправления после 2-го ревью
