+++ app.js
@@ -18,11 +18,13 @@ const errorHandling = require('./middlewares/error');
 
 const corsOptions = require('./middlewares/cors');
 
-const { DB_LINK, PORT = 3000 } = process.env;
+const { DB_LINK_DEV } = require('./utils/const');
+
+const { NODE_ENV, DB_LINK, PORT = 3000 } = process.env;
 
 const app = express();
 
-mongoose.connect(DB_LINK, {
+mongoose.connect(NODE_ENV === 'production' ? DB_LINK : DB_LINK_DEV, {
   useNewUrlParser: true,
 });
 
@@ -30,13 +32,13 @@ app.use(helmet());
 
 app.use(cors(corsOptions));
 
+app.use(requestLogger);
+
 app.use(rateLimiter);
 
 app.use(bodyParser.json());
 
-app.use(requestLogger);
-
-app.use(require('./routes/index'));
+app.use(require('./routes'));
 
 app.use(errorLogger);
